do not dup the scene, let the user handle that
i was trying to keep the user from having to sync rendering while
moving objects around by duplicating the scene every time a picture
was taken.  it seemed like a good idea at first, but we're just lying
to ourselves.  it's the user's responsibility to keep their update()
from messing with their render() anyway, and doing a duplication
without any form of locking on the user side is an invitation for
broken programs.

with this change we go back to the way it was before, and prims only
get copied while they are on the tilers and later being rasterized.
it uses less memory and keeps us from assuming things that are wrong.