add backend Vertex and Primitive types for internal use
this draws a clear distinction between the data processed by the renderer and the one manipulated by the user, decoupling the logic and serving as a first step towards the optimization of geometry storage. there are properties like per-vertex tangents, materials and Vertexattrs (varyings) that have no meaning outside of the rendering pipeline, so it makes no sense to keep space for them.